Move DeboxMap properties into spec
